Fix innerHTML/outerHTML
In IE11, using DOMParser to generate a DOM from a string doesn't include
innerHTML and outerHTML everywhere we need it; it's either that it
doesn't put it on non-HTML tags, or that it doesn't do it for mime type
text/xml, or else a combination of the two. Possible fixes include using
text/html if that would even work with non-HTML tags like <template> and
then stripping the automatically added <html> and <body> tags, falling
back on something other than DOMParser to generate the DOM, and (the
method chosen for this fix) generating innerHTML and outerHTML by
cloning the DOM nodes into a known good dummy HTML node and using that
dummy node's innerHTML (equivalent to the child's outerHTML, simply
strip the outermost tag for its innerHTML).