Fix formatting options array?
The options were being parsed as a space-delimited string, but all the
functions that I could tell were designed to take an individual
option/argument. I'm not sure what the intention was, but what I did was
set it to loop over the array and format the value with each of its
entries, concatenating them into the result, unless the options are
empty in which case call it without the loop or concatenation but don't
send the empty array (the functions expect an undefined value if there
are no options, not an empty array).